Initial support for easy effects v8+ (QT port)
This adds new method for getting last used presets and changes how to parse the output of -p (--presets) option on the new QT version. Getting the last used preset does not work reliably with the new method (at least for flatpak). Need to work together with EE team to find a reliable solution. Also getting the value of global bypass state is disabled until we implement a new solution. Setting it should work though.